Refactor to use pkg and more tablewriter
